move createCatalogProduct into it's own file
